BinarySearchAlgorithm.java
Binary Search Algorithm
This is a Java implementation of the Binary Search Algorithm.

Description
The Binary Search Algorithm is a search algorithm used to find the position of a target value within a sorted array. It works by repeatedly dividing in half the portion of the array that could contain the target value until the value is found or the remaining portion is empty.